Rename Arm SIMD code directories
Future commits will add NEON intrinsics code that will be used to
generate code for both AArch32 and AArch64.
