+++ simd/arm/aarch32/jsimd.c
@@ -1,10 +1,11 @@
 /*
- * jsimd_arm64.c
+ * jsimd_arm.c
  *
  * Copyright 2009 Pierre Ossman <ossman@cendio.se> for Cendio AB
  * Copyright (C) 2011, Nokia Corporation and/or its subsidiary(-ies).
  * Copyright (C) 2009-2011, 2013-2014, 2016, 2018, D. R. Commander.
  * Copyright (C) 2015-2016, 2018, Matthieu Darbois.
+ * Copyright (C) 2019, Google LLC.
  *
  * Based on the x86 SIMD extension for IJG JPEG library,
  * Copyright (C) 1999-2006, MIYASAKA Masaru.
@@ -12,55 +13,48 @@
  *
  * This file contains the interface between the "normal" portions
  * of the library and the SIMD implementations when running on a
- * 64-bit ARM architecture.
+ * 32-bit ARM architecture.
  */
 
 #define JPEG_INTERNALS
-#include "../../jinclude.h"
-#include "../../jpeglib.h"
+#include "../../../jinclude.h"
+#include "../../../jpeglib.h"
+#include "../../../jsimd.h"
+#include "../../../jdct.h"
+#include "../../../jsimddct.h"
 #include "../../jsimd.h"
-#include "../../jdct.h"
-#include "../../jsimddct.h"
-#include "../jsimd.h"
-#include "jconfigint.h"
 
 #include <stdio.h>
 #include <string.h>
 #include <ctype.h>
 
-#define JSIMD_FASTLD3  1
-#define JSIMD_FASTST3  2
-#define JSIMD_FASTTBL  4
-
 static unsigned int simd_support = ~0;
 static unsigned int simd_huffman = 1;
-static unsigned int simd_features = JSIMD_FASTLD3 | JSIMD_FASTST3 |
-                                    JSIMD_FASTTBL;
 
-#if defined(__linux__) || defined(ANDROID) || defined(__ANDROID__)
+#if !defined(__ARM_NEON__) && (defined(__linux__) || defined(ANDROID) || defined(__ANDROID__))
 
 #define SOMEWHAT_SANE_PROC_CPUINFO_SIZE_LIMIT  (1024 * 1024)
 
 LOCAL(int)
-check_cpuinfo(char *buffer, const char *field, char *value)
+check_feature(char *buffer, char *feature)
 {
   char *p;
 
-  if (*value == 0)
+  if (*feature == 0)
     return 0;
-  if (strncmp(buffer, field, strlen(field)) != 0)
+  if (strncmp(buffer, "Features", 8) != 0)
     return 0;
-  buffer += strlen(field);
+  buffer += 8;
   while (isspace(*buffer))
     buffer++;
 
-  /* Check if 'value' is present in the buffer as a separate word */
-  while ((p = strstr(buffer, value))) {
+  /* Check if 'feature' is present in the buffer as a separate word */
+  while ((p = strstr(buffer, feature))) {
     if (p > buffer && !isspace(*(p - 1))) {
       buffer++;
       continue;
     }
-    p += strlen(value);
+    p += strlen(feature);
     if (*p != 0 && !isspace(*p)) {
       buffer++;
       continue;
@@ -76,6 +70,8 @@ parse_proc_cpuinfo(int bufsize)
   char *buffer = (char *)malloc(bufsize);
   FILE *fd;
 
+  simd_support = 0;
+
   if (!buffer)
     return 0;
 
@@ -88,17 +84,8 @@ parse_proc_cpuinfo(int bufsize)
         free(buffer);
         return 0;
       }
-      if (check_cpuinfo(buffer, "CPU part", "0xd03") ||
-          check_cpuinfo(buffer, "CPU part", "0xd07"))
-        /* The Cortex-A53 has a slow tbl implementation.  We can gain a few
-           percent speedup by disabling the use of that instruction.  The
-           speedup on Cortex-A57 is more subtle but still measurable. */
-        simd_features &= ~JSIMD_FASTTBL;
-      else if (check_cpuinfo(buffer, "CPU part", "0x0a1"))
-        /* The SIMD version of Huffman encoding is slower than the C version on
-           Cavium ThunderX.  Also, ld3 and st3 are abyssmally slow on that
-           CPU. */
-        simd_huffman = simd_features = 0;
+      if (check_feature(buffer, "neon"))
+        simd_support |= JSIMD_NEON;
     }
     fclose(fd);
   }
@@ -113,20 +100,13 @@ parse_proc_cpuinfo(int bufsize)
  *
  * FIXME: This code is racy under a multi-threaded environment.
  */
-
-/*
- * ARMv8 architectures support NEON extensions by default.
- * It is no longer optional as it was with ARMv7.
- */
-
-
 LOCAL(void)
 init_simd(void)
 {
 #ifndef NO_GETENV
   char *env = NULL;
 #endif
-#if defined(__linux__) || defined(ANDROID) || defined(__ANDROID__)
+#if !defined(__ARM_NEON__) && (defined(__linux__) || defined(ANDROID) || defined(__ANDROID__))
   int bufsize = 1024; /* an initial guess for the line buffer size limit */
 #endif
 
@@ -135,8 +115,12 @@ init_simd(void)
 
   simd_support = 0;
 
+#if defined(__ARM_NEON__)
   simd_support |= JSIMD_NEON;
-#if defined(__linux__) || defined(ANDROID) || defined(__ANDROID__)
+#elif defined(__linux__) || defined(ANDROID) || defined(__ANDROID__)
+  /* We still have a chance to use NEON regardless of globally used
+   * -mcpu/-mfpu options passed to gcc by performing runtime detection via
+   * /proc/cpuinfo parsing on linux/android */
   while (!parse_proc_cpuinfo(bufsize)) {
     bufsize *= 2;
     if (bufsize > SOMEWHAT_SANE_PROC_CPUINFO_SIZE_LIMIT)
@@ -155,16 +139,6 @@ init_simd(void)
   env = getenv("JSIMD_NOHUFFENC");
   if ((env != NULL) && (strcmp(env, "1") == 0))
     simd_huffman = 0;
-  env = getenv("JSIMD_FASTLD3");
-  if ((env != NULL) && (strcmp(env, "1") == 0))
-    simd_features |= JSIMD_FASTLD3;
-  if ((env != NULL) && (strcmp(env, "0") == 0))
-    simd_features &= ~JSIMD_FASTLD3;
-  env = getenv("JSIMD_FASTST3");
-  if ((env != NULL) && (strcmp(env, "1") == 0))
-    simd_features |= JSIMD_FASTST3;
-  if ((env != NULL) && (strcmp(env, "0") == 0))
-    simd_features &= ~JSIMD_FASTST3;
 #endif
 }
 
@@ -238,20 +212,14 @@ jsimd_rgb_ycc_convert(j_compress_ptr cinfo, JSAMPARRAY input_buf,
 
   switch (cinfo->in_color_space) {
   case JCS_EXT_RGB:
-    if (simd_features & JSIMD_FASTLD3)
-      neonfct = jsimd_extrgb_ycc_convert_neon;
-    else
-      neonfct = jsimd_extrgb_ycc_convert_neon_slowld3;
+    neonfct = jsimd_extrgb_ycc_convert_neon;
     break;
   case JCS_EXT_RGBX:
   case JCS_EXT_RGBA:
     neonfct = jsimd_extrgbx_ycc_convert_neon;
     break;
   case JCS_EXT_BGR:
-    if (simd_features & JSIMD_FASTLD3)
-      neonfct = jsimd_extbgr_ycc_convert_neon;
-    else
-      neonfct = jsimd_extbgr_ycc_convert_neon_slowld3;
+    neonfct = jsimd_extbgr_ycc_convert_neon;
     break;
   case JCS_EXT_BGRX:
   case JCS_EXT_BGRA:
@@ -266,10 +234,7 @@ jsimd_rgb_ycc_convert(j_compress_ptr cinfo, JSAMPARRAY input_buf,
     neonfct = jsimd_extxrgb_ycc_convert_neon;
     break;
   default:
-    if (simd_features & JSIMD_FASTLD3)
-      neonfct = jsimd_extrgb_ycc_convert_neon;
-    else
-      neonfct = jsimd_extrgb_ycc_convert_neon_slowld3;
+    neonfct = jsimd_extrgb_ycc_convert_neon;
     break;
   }
 
@@ -292,20 +257,14 @@ jsimd_ycc_rgb_convert(j_decompress_ptr cinfo, JSAMPIMAGE input_buf,
 
   switch (cinfo->out_color_space) {
   case JCS_EXT_RGB:
-    if (simd_features & JSIMD_FASTST3)
-      neonfct = jsimd_ycc_extrgb_convert_neon;
-    else
-      neonfct = jsimd_ycc_extrgb_convert_neon_slowst3;
+    neonfct = jsimd_ycc_extrgb_convert_neon;
     break;
   case JCS_EXT_RGBX:
   case JCS_EXT_RGBA:
     neonfct = jsimd_ycc_extrgbx_convert_neon;
     break;
   case JCS_EXT_BGR:
-    if (simd_features & JSIMD_FASTST3)
-      neonfct = jsimd_ycc_extbgr_convert_neon;
-    else
-      neonfct = jsimd_ycc_extbgr_convert_neon_slowst3;
+    neonfct = jsimd_ycc_extbgr_convert_neon;
     break;
   case JCS_EXT_BGRX:
   case JCS_EXT_BGRA:
@@ -320,10 +279,7 @@ jsimd_ycc_rgb_convert(j_decompress_ptr cinfo, JSAMPIMAGE input_buf,
     neonfct = jsimd_ycc_extxrgb_convert_neon;
     break;
   default:
-    if (simd_features & JSIMD_FASTST3)
-      neonfct = jsimd_ycc_extrgb_convert_neon;
-    else
-      neonfct = jsimd_ycc_extrgb_convert_neon_slowst3;
+    neonfct = jsimd_ycc_extrgb_convert_neon;
     break;
   }
 
@@ -342,38 +298,12 @@ jsimd_ycc_rgb565_convert(j_decompress_ptr cinfo, JSAMPIMAGE input_buf,
 GLOBAL(int)
 jsimd_can_h2v2_downsample(void)
 {
-  init_simd();
-
-  /* The code is optimised for these values only */
-  if (BITS_IN_JSAMPLE != 8)
-    return 0;
-  if (DCTSIZE != 8)
-    return 0;
-  if (sizeof(JDIMENSION) != 4)
-    return 0;
-
-  if (simd_support & JSIMD_NEON)
-    return 1;
-
   return 0;
 }
 
 GLOBAL(int)
 jsimd_can_h2v1_downsample(void)
 {
-  init_simd();
-
-  /* The code is optimised for these values only */
-  if (BITS_IN_JSAMPLE != 8)
-    return 0;
-  if (DCTSIZE != 8)
-    return 0;
-  if (sizeof(JDIMENSION) != 4)
-    return 0;
-
-  if (simd_support & JSIMD_NEON)
-    return 1;
-
   return 0;
 }
 
@@ -381,18 +311,12 @@ GLOBAL(void)
 jsimd_h2v2_downsample(j_compress_ptr cinfo, jpeg_component_info *compptr,
                       JSAMPARRAY input_data, JSAMPARRAY output_data)
 {
-  jsimd_h2v2_downsample_neon(cinfo->image_width, cinfo->max_v_samp_factor,
-                             compptr->v_samp_factor, compptr->width_in_blocks,
-                             input_data, output_data);
 }
 
 GLOBAL(void)
 jsimd_h2v1_downsample(j_compress_ptr cinfo, jpeg_component_info *compptr,
                       JSAMPARRAY input_data, JSAMPARRAY output_data)
 {
-  jsimd_h2v1_downsample_neon(cinfo->image_width, cinfo->max_v_samp_factor,
-                             compptr->v_samp_factor, compptr->width_in_blocks,
-                             input_data, output_data);
 }
 
 GLOBAL(int)
@@ -428,6 +352,17 @@ jsimd_can_h2v2_fancy_upsample(void)
 GLOBAL(int)
 jsimd_can_h2v1_fancy_upsample(void)
 {
+  init_simd();
+
+  /* The code is optimised for these values only */
+  if (BITS_IN_JSAMPLE != 8)
+    return 0;
+  if (sizeof(JDIMENSION) != 4)
+    return 0;
+
+  if (simd_support & JSIMD_NEON)
+    return 1;
+
   return 0;
 }
 
@@ -441,6 +376,9 @@ GLOBAL(void)
 jsimd_h2v1_fancy_upsample(j_decompress_ptr cinfo, jpeg_component_info *compptr,
                           JSAMPARRAY input_data, JSAMPARRAY *output_data_ptr)
 {
+  jsimd_h2v1_fancy_upsample_neon(cinfo->max_v_samp_factor,
+                                 compptr->downsampled_width, input_data,
+                                 output_data_ptr);
 }
 
 GLOBAL(int)
@@ -510,17 +448,6 @@ jsimd_convsamp_float(JSAMPARRAY sample_data, JDIMENSION start_col,
 GLOBAL(int)
 jsimd_can_fdct_islow(void)
 {
-  init_simd();
-
-  /* The code is optimised for these values only */
-  if (DCTSIZE != 8)
-    return 0;
-  if (sizeof(DCTELEM) != 2)
-    return 0;
-
-  if (simd_support & JSIMD_NEON)
-    return 1;
-
   return 0;
 }
 
@@ -550,7 +477,6 @@ jsimd_can_fdct_float(void)
 GLOBAL(void)
 jsimd_fdct_islow(DCTELEM *data)
 {
-  jsimd_fdct_islow_neon(data);
 }
 
 GLOBAL(void)
@@ -763,29 +689,13 @@ jsimd_huff_encode_one_block(void *state, JOCTET *buffer, JCOEFPTR block,
                             int last_dc_val, c_derived_tbl *dctbl,
                             c_derived_tbl *actbl)
 {
-  if (simd_features & JSIMD_FASTTBL)
-    return jsimd_huff_encode_one_block_neon(state, buffer, block, last_dc_val,
-                                            dctbl, actbl);
-  else
-    return jsimd_huff_encode_one_block_neon_slowtbl(state, buffer, block,
-                                                    last_dc_val, dctbl, actbl);
+  return jsimd_huff_encode_one_block_neon(state, buffer, block, last_dc_val,
+                                          dctbl, actbl);
 }
 
 GLOBAL(int)
 jsimd_can_encode_mcu_AC_first_prepare(void)
 {
-  init_simd();
-
-  if (DCTSIZE != 8)
-    return 0;
-  if (sizeof(JCOEF) != 2)
-    return 0;
-  if (SIZEOF_SIZE_T != 8)
-    return 0;
-
-  if (simd_support & JSIMD_NEON)
-    return 1;
-
   return 0;
 }
 
@@ -794,25 +704,11 @@ jsimd_encode_mcu_AC_first_prepare(const JCOEF *block,
                                   const int *jpeg_natural_order_start, int Sl,
                                   int Al, JCOEF *values, size_t *zerobits)
 {
-  jsimd_encode_mcu_AC_first_prepare_neon(block, jpeg_natural_order_start,
-                                         Sl, Al, values, zerobits);
 }
 
 GLOBAL(int)
 jsimd_can_encode_mcu_AC_refine_prepare(void)
 {
-  init_simd();
-
-  if (DCTSIZE != 8)
-    return 0;
-  if (sizeof(JCOEF) != 2)
-    return 0;
-  if (SIZEOF_SIZE_T != 8)
-    return 0;
-
-  if (simd_support & JSIMD_NEON)
-    return 1;
-
   return 0;
 }
 
@@ -821,7 +717,5 @@ jsimd_encode_mcu_AC_refine_prepare(const JCOEF *block,
                                    const int *jpeg_natural_order_start, int Sl,
                                    int Al, JCOEF *absvalues, size_t *bits)
 {
-  return jsimd_encode_mcu_AC_refine_prepare_neon(block,
-                                                 jpeg_natural_order_start,
-                                                 Sl, Al, absvalues, bits);
+  return 0;
 }
